build-sys: run aclocal --install to populate m4 directory
This work is partly inspired by
https://gitlab.com/grobian/html2text/-/merge_requests/57 where
AC_CONFIG_MACRO_DIRS is used to pick up the local copies from the repo.

Previously the copies in m4 were not used, so rather pointless.  That is
now fixed.  In addition, now run aclocal --install via the release
script prior to a release to refresh the m4 macros we need in configure
so the released code should be completely repeatable with all the
artifacts in the repo.

- drop aclocal.m4 as it is a full copy of all the files in m4
- add missing files in m4/
- drop ACLOCAL_AMFLAGS for autoconf/automake handle inclusion from the
  m4 dir
- refresh m4 macros prior to a release using aclocal --install
- ensure a release commit only contains the changes of configure.ac
  and configure
- use AC_CONFIG_MACRO_DIRS from configure.ac to use local m4 copies

Signed-off-by: Fabian Groffen <grobian@bitzolder.nl>